Split builder helpers and test cases into focused modules
builder.ts had grown a few inline helpers (root-shape extraction,
build-context setup, fragment-definition formatting) and the test
file was a single 1047-line `forEach(['query','mutation'])` block
covering every feature. Both made it harder to locate code by
concept.

- Extract `extractRootShape` into `root-shape.ts`, `createBuildContext`
  (plus the `BuildContext` type) into `build-context.ts`, and
  `formatFragmentDefinitions` (plus the `FragmentDefinition` type)
  into `fragment-definitions.ts`. `serializer.ts` now imports both
  types as types-only.
- Split `builder.test.ts` into themed siblings: basic operations,
  field options (alias/parameters), discriminated-union fragments,
  collections of objects, field types/primitives, and fragment
  hoisting. `builder.test.ts` keeps only the two cases that live
  outside the operation-type loop.
- Move the shared `checkQuery`/`checkError` runners into a new
  `source/test-libraries/check-build-output.ts` (purpose-named, not
  "helpers"). Adjust `tsconfig.sources.json` to exclude
  `test-libraries/`, `tsconfig.unit-tests.json` to include it, and
  `.c8rc.json` to skip the built output from coverage so the folder
  doesn't drag the thresholds.
