+++ .c8rc.json
@@ -14,6 +14,7 @@
     "exclude": [
         "target/build/source/**/*.test.js",
         "target/build/source/**/entry-point.js",
+        "target/build/source/test-libraries/**/*.js",
         "target/build/source/zod-graphql-client/operation-result.js"
     ]
 }
