+++ source/tsconfig.sources.json
@@ -7,5 +7,5 @@
         "skipLibCheck": true
     },
     "include": ["./**/*.ts"],
-    "exclude": ["./**/*.test.ts"]
+    "exclude": ["./**/*.test.ts", "./test-libraries/**/*.ts"]
 }
