+++ source/tsconfig.unit-tests.json
@@ -7,5 +7,5 @@
         "skipLibCheck": true
     },
     "references": [{ "path": "./tsconfig.sources.json" }],
-    "include": ["./**/*.test.ts"]
+    "include": ["./**/*.test.ts", "./test-libraries/**/*.ts"]
 }
